+++ packages/python/archil/openapi/models/aws_sts_user_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class AwsStsUserType(str, Enum):
-    AWSSTS = "awssts"
-
-    def __str__(self) -> str:
-        return str(self.value)