+++ packages/python/archil/openapi/models/azure_blob_storage_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class AzureBlobStorageType(str, Enum):
-    AZURE_BLOB = "azure-blob"
-
-    def __str__(self) -> str:
-        return str(self.value)