+++ packages/python/archil/openapi/models/cloudflare_r2_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class CloudflareR2Type(str, Enum):
-    R2 = "r2"
-
-    def __str__(self) -> str:
-        return str(self.value)