+++ packages/python/archil/openapi/models/disk_response_capabilities_item.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class DiskResponseCapabilitiesItem(str, Enum):
-    CHECKPOINTS = "checkpoints"
-
-    def __str__(self) -> str:
-        return str(self.value)