+++ packages/python/archil/openapi/models/google_cloud_storage_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class GoogleCloudStorageType(str, Enum):
-    GCS = "gcs"
-
-    def __str__(self) -> str:
-        return str(self.value)