+++ packages/python/archil/openapi/models/mount_response_access_mode.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class MountResponseAccessMode(str, Enum):
-    RO = "ro"
-    RW = "rw"
-
-    def __str__(self) -> str:
-        return str(self.value)