+++ packages/python/archil/openapi/models/mount_response_connection_status.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class MountResponseConnectionStatus(str, Enum):
-    CONNECTED = "connected"
-    DISCONNECTED = "disconnected"
-
-    def __str__(self) -> str:
-        return str(self.value)