+++ packages/python/archil/openapi/models/remove_disk_user_user_type.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class RemoveDiskUserUserType(str, Enum):
-    AWSSTS = "awssts"
-    TOKEN = "token"
-
-    def __str__(self) -> str:
-        return str(self.value)