+++ packages/python/archil/openapi/models/sandbox_port_mapping_protocol.py
@@ -1,9 +1,0 @@
-from enum import Enum
-
-
-class SandboxPortMappingProtocol(str, Enum):
-    TCP = "tcp"
-    UDP = "udp"
-
-    def __str__(self) -> str:
-        return str(self.value)