+++ packages/python/archil/openapi/models/token_user_type.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class TokenUserType(str, Enum):
-    TOKEN = "token"
-
-    def __str__(self) -> str:
-        return str(self.value)